refactored animations, added color, renamed shared folder, moved animation packages to devDependencies
