+++ components/HomePage/FeaturedPost.jsx
@@ -1,4 +1,4 @@
-import List from '../shared/List';
+import List from '../Shared/List';
 const FeaturedPost = ({ blogs }) => {
 	return (
 		<section className='m-6 space-y-6'>
