+++ package.json
@@ -10,11 +10,9 @@
   },
   "dependencies": {
     "emailjs-com": "^3.2.0",
-    "framer-motion": "^4.1.17",
     "next": "11.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-intersection-observer": "^8.32.0",
     "react-markdown": "^6.0.2",
     "react-syntax-highlighter": "^15.4.3"
   },
@@ -24,8 +22,10 @@
     "autoprefixer": "^10.2.6",
     "eslint": "7.29.0",
     "eslint-config-next": "11.0.1",
+    "framer-motion": "^4.1.17",
     "next-seo": "^4.26.0",
     "postcss": "^8.3.5",
+    "react-intersection-observer": "^8.32.0",
     "rehype-raw": "^5.1.0",
     "rehype-sanitize": "^4.0.0",
     "remark-gfm": "^1.0.0",
