+++ pages/blogs/index.js
@@ -23,5 +23,6 @@ export async function getStaticProps() {
 		props: {
 			blog: data,
 		},
+		revalidate: 86400,
 	};
 }
